商品削除機能編集１

+++ app/controllers/items_controller.rb
@@ -41,7 +41,7 @@ class ItemsController < ApplicationController
 
   def destroy
     item = Item.find(params[:id])
-    item.destroy
+    item.destroy if current_user.id == item.user_id
     redirect_to root_path 
   end
